+++ tests/format.test.js
@@ -1,26 +1,4 @@
-const prettier = require('prettier')
-const path = require('path')
-const fs = require('fs')
-const { exec } = require('child_process')
 const { t, yes, no, format } = require('./utils')
-const { promisify } = require('util')
-const execAsync = promisify(exec)
-
-async function formatFixture(name, extension) {
-  let binPath = path.resolve(__dirname, '../node_modules/.bin/prettier')
-  let filePath = path.resolve(__dirname, `fixtures/${name}/index.${extension}`)
-  let pluginPath = path.resolve(__dirname, '../dist/index.js')
-
-  let cmd
-
-  if (prettier.version.startsWith('2.')) {
-    cmd = `${binPath} ${filePath} --plugin-search-dir ${__dirname} --plugin ${pluginPath}`
-  } else {
-    cmd = `${binPath} ${filePath} --plugin ${pluginPath}`
-  }
-
-  return execAsync(cmd).then(({ stdout }) => stdout.trim())
-}
 
 let html = [
   t`<div class="${yes}"></div>`,
@@ -160,119 +138,37 @@ let tests = {
     t`<div [ngClass]="[someVar ?? '${yes}']"></div>`,
     t`<div [ngClass]="{ '${yes}': true }"></div>`,
     t`<div [ngClass]="clsx('${yes}')"></div>`,
-    prettier.version.startsWith('2.')
-    ? [
-      `<div [ngClass]="{ 'sm:p-0 p-0': (some.thing | urlPipe: { option: true } | async), 'sm:p-0 p-0': true }"></div>`,
-      `<div [ngClass]="{ 'p-0 sm:p-0': (some.thing | urlPipe : { option: true } | async), 'p-0 sm:p-0': true }"></div>`,
-    ]
-    : t`<div [ngClass]="{ '${yes}': (some.thing | urlPipe: { option: true } | async), '${yes}': true }"></div>`,
+    t`<div [ngClass]="{ '${yes}': (some.thing | urlPipe: { option: true } | async), '${yes}': true }"></div>`,
     t`<div [ngClass]="{ '${yes}': foo && bar?.['baz'] }" class="${yes}"></div>`,
 
     // TODO: Enable this test — it causes console noise but not a failure
     // t`<div [ngClass]="{ '${no}': foo && definitely&a:syntax*error }" class="${yes}"></div>`,
   ],
   css: [...css, t`@apply ${yes} !important;`],
-  scss: [...css, t`@apply ${yes} #{!important};`],
+  scss: [
+    ...css,
+    t`@apply ${yes} #{!important};`,
+    t`@apply ${yes} #{'!important'};`,
+    t`@apply ${yes} #{"!important"};`,
+
+    // These shouldn't ever be used but they are valid
+    // syntax so we might as well not break them
+    t`@apply ${yes} #{""!important""};`,
+    t`@apply ${yes} #{'''!important'''};`,
+    t`@apply ${yes} #{"'"'"!important"'"'"};`,
+  ],
   less: [...css, t`@apply ${yes} !important;`],
   babel: javascript,
   typescript: javascript,
   'babel-ts': javascript,
   flow: javascript,
   'babel-flow': javascript,
-  ...(
-    prettier.version.startsWith('2.')
-      ? { espree: javascript }
-      : { acorn: javascript }
-  ),
+  acorn: javascript,
   meriyah: javascript,
   mdx: javascript
     .filter((test) => !test.find((t) => /^\/\*/.test(t)))
     .map((test) => test.map((t) => t.replace(/^;/, ''))),
 }
-
-let fixtures = [
-  {
-    name: 'no prettier config',
-    dir: 'no-prettier-config',
-    output: '<div class="bg-red-500 sm:bg-tomato"></div>',
-  },
-  {
-    name: 'inferred config path',
-    dir: 'basic',
-    output: '<div class="bg-red-500 sm:bg-tomato"></div>',
-  },
-  {
-    name: 'inferred config path (.cjs)',
-    dir: 'cjs',
-    output: '<div class="bg-red-500 sm:bg-hotpink"></div>',
-  },
-  {
-    name: 'using esm config',
-    dir: 'esm',
-    output: '<div class="bg-red-500 sm:bg-hotpink"></div>',
-  },
-  {
-    name: 'using esm config (explicit path)',
-    dir: 'esm-explicit',
-    output: '<div class="bg-red-500 sm:bg-hotpink"></div>',
-  },
-  {
-    name: 'using ts config',
-    dir: 'ts',
-    output: '<div class="bg-red-500 sm:bg-hotpink"></div>',
-  },
-  {
-    name: 'using ts config (explicit path)',
-    dir: 'ts-explicit',
-    output: '<div class="bg-red-500 sm:bg-hotpink"></div>',
-  },
-  {
-    name: 'using v3.2.7',
-    dir: 'v3-2',
-    output: '<div class="bg-red-500 sm:bg-tomato"></div>',
-  },
-  {
-    name: 'plugins',
-    dir: 'plugins',
-    output: '<div class="uppercase foo sm:bar"></div>',
-  },
-  {
-    name: 'customizations: js/jsx',
-    dir: 'custom-jsx',
-    ext: 'jsx',
-    output: `const a = sortMeFn("p-2 sm:p-1");
-const b = sortMeFn({
-  foo: "p-2 sm:p-1",
-});
-
-const c = dontSortFn("sm:p-1 p-2");
-const d = sortMeTemplate\`p-2 sm:p-1\`;
-const e = dontSortMeTemplate\`sm:p-1 p-2\`;
-const f = tw.foo\`p-2 sm:p-1\`;
-const g = tw.foo.bar\`p-2 sm:p-1\`;
-const h = no.foo\`sm:p-1 p-2\`;
-const i = no.tw\`sm:p-1 p-2\`;
-
-const A = (props) => <div className={props.sortMe} />;
-const B = () => <A sortMe="p-2 sm:p-1" dontSort="sm:p-1 p-2" />;`,
-  },
-  {
-    name: 'customizations: vue',
-    dir: 'custom-vue',
-    ext: 'vue',
-    output: `<script setup>
-let a = sortMeFn("p-2 sm:p-1");
-let b = sortMeFn({ "p-2 sm:p-1": true });
-let c = dontSortFn("sm:p-1 p-2");
-let d = sortMeTemplate\`p-2 sm:p-1\`;
-let e = dontSortMeTemplate\`sm:p-1 p-2\`;
-</script>
-<template>
-  <div class="p-2 sm:p-1" sortMe="p-2 sm:p-1" dontSortMe="sm:p-1 p-2"></div>
-  <div :class="{ 'p-2 sm:p-1': true }"></div>
-</template>`,
-  },
-]
 
 describe('parsers', () => {
   for (let parser in tests) {
@@ -293,47 +189,9 @@ describe('other', () => {
 
   test('parasite utilities', async () => {
     expect(
-      await format('<div class="group peer unknown-class p-0 container"></div>'),
+      await format(
+        '<div class="group peer unknown-class p-0 container"></div>',
+      ),
     ).toEqual('<div class="unknown-class group peer container p-0"></div>')
   })
-
-  test('explicit config path', async () => {
-    expect(
-      await format('<div class="sm:bg-tomato bg-red-500"></div>', {
-        tailwindConfig: path.resolve(
-          __dirname,
-          'fixtures/basic/tailwind.config.js',
-        ),
-      }),
-    ).toEqual('<div class="bg-red-500 sm:bg-tomato"></div>')
-  })
-})
-
-describe('fixtures', () => {
-  let configs = [
-    {
-      from: __dirname + '/../.prettierignore',
-      to: __dirname + '/../.prettierignore.testing',
-    },
-    {
-      from: __dirname + '/../prettier.config.js',
-      to: __dirname + '/../prettier.config.js.testing',
-    },
-  ]
-
-  // Temporarily move config files out of the way so they don't interfere with the tests
-  beforeAll(() =>
-    Promise.all(configs.map(({ from, to }) => fs.promises.rename(from, to))),
-  )
-
-  afterAll(() =>
-    Promise.all(configs.map(({ from, to }) => fs.promises.rename(to, from))),
-  )
-
-  for (const fixture of fixtures) {
-    test(fixture.name, async () => {
-      let formatted = await formatFixture(fixture.dir, fixture.ext ?? 'html')
-      expect(formatted).toEqual(fixture.output)
-    })
-  }
 })