+++ tests/utils.js
@@ -26,6 +26,10 @@ module.exports.t = function t(strings, ...values) {
   return [input, output]
 }
 
+let pluginPath = path.resolve(__dirname, '../dist/index.mjs')
+
+module.exports.pluginPath = pluginPath
+
 module.exports.format = async function format(str, options = {}) {
   let result = await prettier.format(str, {
     pluginSearchDirs: [__dirname], // disable plugin autoload
@@ -34,10 +38,7 @@ module.exports.format = async function format(str, options = {}) {
     printWidth: 9999,
     parser: 'html',
     ...options,
-    plugins: [
-      ...options.plugins ?? [],
-      path.resolve(__dirname, '../dist/index.js')
-    ],
+    plugins: [...(options.plugins ?? []), pluginPath],
   })
 
   return result.trim()